Ruff 0.16 (#219)
* Bump ruff to 0.16

* Fix a first round of new ruff linting issues

* Fix G201 linting issue

* Fix TRY401 exceptions

* Fix UP040 linting issue

* Fix SIM102 linting issue

* Fix TRY002 linting issue

* Fix DTZ*** linting issues

* Fix SIM115 linting issues

* Fix C408 linting issues

* Fix LOG014 linting issues

* Fix various linting issues

* Address review remarks

+++ qtoggleserver/core/expressions/functions/rounding.py
@@ -11,7 +11,7 @@ class FloorFunction(Function):
     async def _eval(self, context: EvalContext) -> EvalResult:
         eval_args = await self.eval_args(context)
 
-        return int(math.floor(eval_args[0]))
+        return math.floor(eval_args[0])
 
 
 @function("CEIL")
@@ -21,7 +21,7 @@ class CeilFunction(Function):
     async def _eval(self, context: EvalContext) -> EvalResult:
         eval_args = await self.eval_args(context)
 
-        return int(math.ceil(eval_args[0]))
+        return math.ceil(eval_args[0])
 
 
 @function("ROUND")
